Remove deprecated utils and clean up types
